requirements.txt
openai 
streamlit
google.generativeai